+++ src/blockchain/SmartContract.ts
@@ -94,6 +94,8 @@ export type SmartContractTransaction = Transaction & {
     blockchainLogs: string
     vmLogs: string
     debugLogs: string
+    oldStorage?: Cell
+    newStorage?: Cell
 }
 export type MessageParams = Partial<{
     now: number,
@@ -140,8 +142,27 @@ export class TimeError extends Error {
 }
 
 export class EmulationError extends Error {
-    constructor(public error: string, public vmLogs?: string, public exitCode?: number) {
-        super(`Error while executing transaction: ${error}`)
+    constructor(
+        public error: string,
+        public vmLogs?: string,
+        public exitCode?: number,
+        public blockchainLogs?: string,
+        public debugLogs?: string,
+    ) {
+        let errMsg = `Error while executing transaction: ${error}`
+        if (exitCode !== undefined) {
+            errMsg += `\nExit code: ${exitCode}`
+        }
+        if (vmLogs !== undefined) {
+            errMsg += `\nVM logs:\n${vmLogs}`
+        }
+        if (blockchainLogs !== undefined) {
+            errMsg += `\nBlockchain logs:\n${blockchainLogs}`
+        }
+        if (debugLogs !== undefined) {
+            errMsg += `\nDebug logs:\n${debugLogs}`
+        }
+        super(errMsg)
     }
 }
 
@@ -253,34 +274,45 @@ export class SmartContract {
         }
     }
 
-    receiveMessage(message: Message, params?: MessageParams) {
+    async receiveMessage(message: Message, params?: MessageParams) {
         // Sync now with blockchain instance if not specified in parameters
         params = {
             now: this.blockchain.now,
             ...params,
         }
-        return this.runCommon(() => this.blockchain.executor.runTransaction({
+        return await this.runCommon(() => this.blockchain.executor.runTransaction({
             ...this.createCommonArgs(params),
             message: beginCell().store(storeMessage(message)).endCell(),
         }))
     }
 
-    runTickTock(which: TickOrTock, params?: MessageParams) {
-        return this.runCommon(() => this.blockchain.executor.runTickTock({
+    async runTickTock(which: TickOrTock, params?: MessageParams) {
+        return await this.runCommon(() => this.blockchain.executor.runTickTock({
             ...this.createCommonArgs(params),
             which,
         }))
     }
 
-    protected runCommon(run: () => EmulationResult): SmartContractTransaction {
-        const res = run()
+    protected async runCommon(run: () => Promise<EmulationResult>): Promise<SmartContractTransaction> {
+        let oldStorage: Cell | undefined = undefined
+        if (this.blockchain.recordStorage && this.account.account?.storage.state.type === 'active') {
+            oldStorage = this.account.account?.storage.state.state.data ?? undefined
+        }
+
+        const res = await run()
 
         if (this.verbosity.print && this.verbosity.blockchainLogs && res.logs.length > 0) {
             console.log(res.logs)
         }
 
         if (!res.result.success) {
-            throw new EmulationError(res.result.error, res.result.vmResults?.vmLog, res.result.vmResults?.vmExitCode)
+            throw new EmulationError(
+                res.result.error,
+                res.result.vmResults?.vmLog,
+                res.result.vmResults?.vmExitCode,
+                res.logs.length === 0 ? undefined : res.logs,
+                res.debugLogs.length === 0 ? undefined : res.debugLogs,
+            )
         }
 
         if (this.verbosity.print && this.verbosity.vmLogs !== 'none' && res.result.vmLog.length > 0) {
@@ -297,20 +329,27 @@ export class SmartContract {
         this.#parsedAccount = undefined
         this.#lastTxTime = tx.now
 
+        let newStorage: Cell | undefined = undefined
+        if (this.blockchain.recordStorage && this.account.account?.storage.state.type === 'active') {
+            newStorage = this.account.account?.storage.state.state.data ?? undefined
+        }
+
         return {
             ...tx,
             blockchainLogs: res.logs,
             vmLogs: res.result.vmLog,
             debugLogs: res.debugLogs,
+            oldStorage,
+            newStorage,
         }
     }
 
-    get(method: string | number, stack: TupleItem[] = [], params?: GetMethodParams): GetMethodResult {
+    async get(method: string | number, stack: TupleItem[] = [], params?: GetMethodParams): Promise<GetMethodResult> {
         if (this.account.account?.storage.state.type !== 'active') {
             throw new Error('Trying to run get method on non-active contract')
         }
 
-        const res = this.blockchain.executor.runGetMethod({
+        const res = await this.blockchain.executor.runGetMethod({
             code: this.account.account?.storage.state.state.code!,
             data: this.account.account?.storage.state.state.data!,
             methodId: typeof method === 'string' ? getSelectorForMethod(method) : method,