+++ src/executor/Executor.ts
@@ -205,7 +205,13 @@ class Heap {
     }
 }
 
-export class Executor {
+export interface IExecutor {
+    runGetMethod(args: GetMethodArgs): Promise<GetMethodResult>
+    runTickTock(args: RunTickTockArgs): Promise<EmulationResult>
+    runTransaction(args: RunTransactionArgs): Promise<EmulationResult>
+}
+
+export class Executor implements IExecutor {
     private module: any
     private heap: Heap
     private emulator?: {
@@ -228,7 +234,7 @@ export class Executor {
         return ex
     }
 
-    runGetMethod(args: GetMethodArgs): GetMethodResult {
+    async runGetMethod(args: GetMethodArgs): Promise<GetMethodResult> {
         const params: GetMethodInternalParams = {
             code: args.code.toBoc().toString('base64'),
             data: args.data.toBoc().toString('base64'),
@@ -267,7 +273,7 @@ export class Executor {
 
     private runCommon(args: (string | number)[]): EmulationResult {
         this.debugLogs = []
-        const resp = JSON.parse(this.extractString(this.invoke('_emulate', args)))
+        const resp = JSON.parse(this.extractString(this.invoke('_emulate_with_emulator', args)))
         const debugLogs = this.debugLogs.join('\n')
 
         if (resp.fail) {
@@ -299,7 +305,7 @@ export class Executor {
         };
     }
 
-    runTickTock(args: RunTickTockArgs): EmulationResult {
+    async runTickTock(args: RunTickTockArgs): Promise<EmulationResult> {
         const params: EmulationInternalParams = {
             ...runCommonArgsToInternalParams(args),
             is_tick_tock: true,
@@ -315,7 +321,7 @@ export class Executor {
         ])
     }
 
-    runTransaction(args: RunTransactionArgs): EmulationResult {
+    async runTransaction(args: RunTransactionArgs): Promise<EmulationResult> {
         const params: EmulationInternalParams = runCommonArgsToInternalParams(args)
 
         return this.runCommon([