+++ src/index.ts
@@ -5,6 +5,7 @@ export {
 
 export {
     Blockchain,
+    toSandboxContract,
     SendMessageResult,
     BlockchainTransaction,
     PendingMessage,
@@ -50,6 +51,14 @@ export {
 
 export {
     TickOrTock,
+    IExecutor,
+    Executor,
+    GetMethodArgs as ExecutorGetMethodArgs,
+    GetMethodResult as ExecutorGetMethodResult,
+    RunTickTockArgs as ExecutorRunTickTockArgs,
+    EmulationResult as ExecutorEmulationResult,
+    RunTransactionArgs as ExecutorRunTransactionArgs,
+    ExecutorVerbosity,
 } from './executor/Executor';
 
 export {