+++ src/treasury/Treasury.ts
@@ -1,4 +1,21 @@
-import { Address, beginCell, Cell, Contract, contractAddress, ContractProvider, Dictionary, DictionaryValue, internal, loadMessageRelaxed, MessageRelaxed, Sender, SenderArguments, SendMode, storeMessageRelaxed } from "@ton/core";
+import {
+    Address,
+    beginCell,
+    Cell,
+    Contract,
+    contractAddress,
+    ContractProvider,
+    Dictionary,
+    DictionaryValue,
+    internal,
+    loadMessageRelaxed,
+    MessageRelaxed,
+    Sender,
+    SenderArguments,
+    SendMode,
+    StateInit,
+    storeMessageRelaxed
+} from "@ton/core";
 
 const DictionaryMessageValue: DictionaryValue<{ sendMode: SendMode, message: MessageRelaxed }> = {
     serialize(src, builder) {
@@ -26,6 +43,9 @@ function senderArgsToMessageRelaxed(args: SenderArguments): MessageRelaxed {
     })
 }
 
+/**
+ * @class TreasuryContract is Wallet v4 alternative. For additional information see {@link Blockchain.treasury}
+ */
 export class TreasuryContract implements Contract {
     static readonly code = Cell.fromBase64('te6cckEBBAEARQABFP8A9KQT9LzyyAsBAgEgAwIAWvLT/+1E0NP/0RK68qL0BNH4AH+OFiGAEPR4b6UgmALTB9QwAfsAkTLiAbPmWwAE0jD+omUe')
 
@@ -34,7 +54,7 @@ export class TreasuryContract implements Contract {
     }
 
     readonly address: Address;
-    readonly init: { code: Cell, data: Cell };
+    readonly init: StateInit;
     readonly subwalletId: bigint;
 
     constructor(workchain: number, subwalletId: bigint) {
@@ -46,6 +66,11 @@ export class TreasuryContract implements Contract {
         this.subwalletId = subwalletId;
     }
 
+    /**
+     * Send bulk messages using one external message.
+     * @param messages Messages to send
+     * @param sendMode Send mode of every message
+     */
     async sendMessages(provider: ContractProvider, messages: MessageRelaxed[], sendMode?: SendMode) {
         let transfer = this.createTransfer({
             sendMode: sendMode,
@@ -54,10 +79,16 @@ export class TreasuryContract implements Contract {
         await provider.external(transfer)
     }
 
+    /**
+     * Sends message by arguments specified.
+     */
     async send(provider: ContractProvider, args: SenderArguments) {
         await this.sendMessages(provider, [senderArgsToMessageRelaxed(args)], args.sendMode ?? undefined)
     }
 
+    /**
+     * @returns Sender
+     */
     getSender(provider: ContractProvider): Treasury {
         return {
             address: this.address,
@@ -71,10 +102,16 @@ export class TreasuryContract implements Contract {
         };
     }
 
+    /**
+     * @returns wallet balance in nanoTONs
+     */
     async getBalance(provider: ContractProvider): Promise<bigint> {
         return (await provider.getState()).balance
     }
 
+    /**
+     * Creates transfer cell for {@link sendMessages}.
+     */
     createTransfer(args: {
         messages: MessageRelaxed[]
         sendMode?: SendMode,